Move fetchContributors method to ContributorList component because it is used by only ContributorList component
